Removed details about release process
Why you made the change:

I did this because the majority of the community doesn't care about the details
about how we mechanically make releases. Hence, it just seems like noise. If we
start to do something unique in terms of our process then maybe it would make
sense to write a blog post about it, or reference it in a wiki or something.
But, until then I don't think it is really worth having around.